requirements.txt
<<<<<<< HEAD
aiohttp==3.7.4.post0
apiclient==1.0.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
discord.py==1.7.3
docopt==0.6.2
Flask==2.0.1
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
monotonic==1.6
multidict==5.1.0
numpy==1.21.0
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
peewee==3.14.4
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodomex==3.10.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-valve==0.2.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
steam==1.2.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
vdf==3.4
vk-api==11.9.4
Werkzeug==2.0.1
yarl==1.6.3
=======
aiohttp==3.7.4.post0
apiclient==1.0.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
discord.py==1.7.3
docopt==0.6.2
Flask==2.0.1
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
monotonic==1.6
multidict==5.1.0
numpy==1.21.0
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
peewee==3.14.4
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodomex==3.10.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-valve==0.2.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
steam==1.2.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
vdf==3.4
vk-api==11.9.4
Werkzeug==2.0.1
yarl==1.6.3
>>>>>>> 75b13045fc71e61cb736d610ba0be5d6d89e8926
